Add a #define NOMINMAX at the start of source files that fail to build in GHA but build locally. I don't know why this would help but that's the best idea I have right now. This commit also includes an additional paremter on onLoadError - the error URL - since I am also debugging an apparent failure to load the SL login page reliably

+++ src/dullahan_impl.cpp
@@ -167,7 +167,7 @@ std::string convert_wide_to_string(const wchar_t* in, unsigned int code_page)
     std::string out;
     if (in)
     {
-        int len_in = wcslen(in);
+        int len_in = (int)wcslen(in);
         int len_out = WideCharToMultiByte(
             code_page,
             0,
